+++ cognee/cognee_skills/example_skills/data-extraction/SKILL.md
