+++ cognee/cognee_skills/example_skills/data-extraction/references/schemas.md
