+++ cognee/cognee_skills/example_skills/summarize/SKILL.md
