+++ cognee/cognee_skills/example_skills/summarize/references/prompts.md
