+++ cognee/cognee_skills/tests/test_inspect.py
@@ -1,11 +1,11 @@
-"""Tests for cognee.skills.inspect — unit tests with mocked LLM + graph."""
+"""Tests for cognee.cognee_skills.inspect — unit tests with mocked LLM + graph."""
 
 import asyncio
 import unittest
 from unittest.mock import AsyncMock, patch, MagicMock
 from uuid import uuid4
 
-from cognee.skills.models.skill_inspection import InspectionResult
+from cognee.cognee_skills.models.skill_inspection import InspectionResult
 
 
 def _make_skill_node(skill_id="test-skill", name="Test Skill"):
@@ -45,10 +45,10 @@ MOCK_INSPECTION_RESULT = InspectionResult(
 
 
 class TestInspectSkill(unittest.TestCase):
-    @patch("cognee.skills.inspect.add_data_points", new_callable=AsyncMock)
-    @patch("cognee.skills.inspect.get_llm_config")
-    @patch("cognee.skills.inspect.LLMGateway.acreate_structured_output", new_callable=AsyncMock)
-    @patch("cognee.skills.inspect.get_graph_engine", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.add_data_points", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_llm_config")
+    @patch("cognee.cognee_skills.inspect.LLMGateway.acreate_structured_output", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_graph_engine", new_callable=AsyncMock)
     def test_inspect_returns_inspection(self, mock_engine_fn, mock_llm, mock_config, mock_add_dp):
         skill_node = _make_skill_node()
         run1 = _make_failed_run(score=0.1)
@@ -60,7 +60,7 @@ class TestInspectSkill(unittest.TestCase):
         mock_llm.return_value = MOCK_INSPECTION_RESULT
         mock_config.return_value = MagicMock(llm_model="openai/gpt-4o-mini")
 
-        from cognee.skills.inspect import inspect_skill
+        from cognee.cognee_skills.inspect import inspect_skill
 
         result = asyncio.run(inspect_skill("test-skill", min_runs=1))
 
@@ -73,7 +73,7 @@ class TestInspectSkill(unittest.TestCase):
         assert result.inspection_confidence == 0.85
         mock_add_dp.assert_called_once()
 
-    @patch("cognee.skills.inspect.get_graph_engine", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_graph_engine", new_callable=AsyncMock)
     def test_inspect_no_failed_runs(self, mock_engine_fn):
         skill_node = _make_skill_node()
 
@@ -81,16 +81,16 @@ class TestInspectSkill(unittest.TestCase):
         engine.get_nodeset_subgraph = AsyncMock(return_value=([skill_node], []))
         mock_engine_fn.return_value = engine
 
-        from cognee.skills.inspect import inspect_skill
+        from cognee.cognee_skills.inspect import inspect_skill
 
         result = asyncio.run(inspect_skill("test-skill", min_runs=1))
 
         assert result is None
 
-    @patch("cognee.skills.inspect.add_data_points", new_callable=AsyncMock)
-    @patch("cognee.skills.inspect.get_llm_config")
-    @patch("cognee.skills.inspect.LLMGateway.acreate_structured_output", new_callable=AsyncMock)
-    @patch("cognee.skills.inspect.get_graph_engine", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.add_data_points", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_llm_config")
+    @patch("cognee.cognee_skills.inspect.LLMGateway.acreate_structured_output", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_graph_engine", new_callable=AsyncMock)
     def test_inspect_aggregates_multiple_runs(
         self, mock_engine_fn, mock_llm, mock_config, mock_add_dp
     ):
@@ -103,7 +103,7 @@ class TestInspectSkill(unittest.TestCase):
         mock_llm.return_value = MOCK_INSPECTION_RESULT
         mock_config.return_value = MagicMock(llm_model="openai/gpt-4o-mini")
 
-        from cognee.skills.inspect import inspect_skill
+        from cognee.cognee_skills.inspect import inspect_skill
 
         result = asyncio.run(inspect_skill("test-skill", min_runs=1))
 
@@ -111,7 +111,7 @@ class TestInspectSkill(unittest.TestCase):
         assert result.analyzed_run_count == 5
         assert len(result.analyzed_run_ids) == 5
 
-    @patch("cognee.skills.inspect.get_graph_engine", new_callable=AsyncMock)
+    @patch("cognee.cognee_skills.inspect.get_graph_engine", new_callable=AsyncMock)
     def test_inspect_insufficient_runs(self, mock_engine_fn):
         skill_node = _make_skill_node()
         run1 = _make_failed_run(score=0.2)
@@ -120,7 +120,7 @@ class TestInspectSkill(unittest.TestCase):
         engine.get_nodeset_subgraph = AsyncMock(return_value=([skill_node, run1], []))
         mock_engine_fn.return_value = engine
 
-        from cognee.skills.inspect import inspect_skill
+        from cognee.cognee_skills.inspect import inspect_skill
 
         # Requires 3 runs but only 1 exists
         result = asyncio.run(inspect_skill("test-skill", min_runs=3))
