+++ cognee/cognee_skills/tests/test_parser.py
@@ -5,7 +5,7 @@ import tempfile
 
 import pytest
 
-from cognee.skills.parser.skill_parser import (
+from cognee.cognee_skills.parser.skill_parser import (
     parse_skill_file,
     parse_skill_folder,
     parse_skills_folder,
